Fix type of selectivity estimate in responses.
fix #366

+++ arangodb-net-standard/AqlFunctionApi/Models/PostExplainAqlQueryResponseIndex.cs
@@ -10,7 +10,7 @@ namespace ArangoDBNetStandard.AqlFunctionApi.Models
         public string Type { get; set; }
         public string Name { get; set; }
         public IEnumerable<string> Fields { get; set; }
-        public int? SelectivityEstimate { get; set; }
+        public double? SelectivityEstimate { get; set; }
         public bool? Unique { get; set; }
         public bool? Sparse { get; set; }
         public bool? Deduplicate { get; set; }
